Add few functions
The len() function added, to count the characters in a string.
The get_path() function added, to get the path string
The com_exists() function added, to check if a command exists.
command may be an absolue path such as: /bin/ls, or a simple string such as ls

The main.h and shell.c files are updated to reflect the addition of
the new functions.